requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Database Drivers (choose based on your database)
# For MySQL/MariaDB
pymysql==1.1.0
cryptography==41.0.7

# For PostgreSQL
# psycopg2-binary==2.9.9

# For SQL Server
# pyodbc==5.0.1
# or
# pymssql==2.2.11

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Templates
jinja2==3.1.3

# Google Forms API (optional, for advanced prefilling)
google-api-python-client==2.111.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utilities
python-dateutil==2.8.2
PyJWT

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
